style: 코드 정리

+++ src/main/java/com/web/billim/product/repository/ImageProductRepository.java
@@ -3,5 +3,5 @@ package com.web.billim.product.repository;
 import com.web.billim.product.domain.ImageProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageProductRepository extends JpaRepository<ImageProduct, Integer> {
+public interface ImageProductRepository extends JpaRepository<ImageProduct, Long> {
 }
